+++ Reflection/src/utopia/reflection/component/drawing/immutable/TextDrawer2.scala
@@ -2,6 +2,7 @@ package utopia.reflection.component.drawing.immutable
 
 import utopia.genesis.graphics.MeasuredText
 import utopia.paradigm.color.Color
+import utopia.paradigm.enumeration.Alignment
 import utopia.reflection.component.drawing.template.DrawLevel.Normal
 import utopia.reflection.component.drawing.template.{DrawLevel, TextDrawerLike2}
 import utopia.reflection.shape.stack.StackInsets
@@ -14,5 +15,5 @@ import utopia.reflection.text.Font
   */
 case class TextDrawer2(override val text: MeasuredText, override val font: Font,
                        override val insets: StackInsets = StackInsets.any, override val color: Color = Color.textBlack,
-                       override val drawLevel: DrawLevel = Normal)
+                       override val alignment: Alignment = Alignment.Left, override val drawLevel: DrawLevel = Normal)
 	extends TextDrawerLike2